all: engine only updating state based on raft log
Further operations are always executed based on the current state. This
includes manipulating VIPs and ipvs state in kernel.

Changes are always made based on diffs between the current state in
memory and the real state committed to the machine's kernel. Rules are
then added, removed or updated as needed.

It's possible to improve this code even more in the future to avoid
unnecessary update calls to the ipvs stack.